Added first round of styling to index.html
Pretty badly styled index page has had its first round of styling, I am trying to figure out how bootstrap makes it look so good and I am hoping there is a design section
to the course or the styling comes with practice and knowledge. I though about using bootstrap but felt it would be better to stay true to the course.