Styles and structure improvements for index
I wasn't getting what I wanted with the unordered list so I split it up into divs. This may not be the
last time I add styles to it because its far from good. There are still slivers of white below the images
and the general theme of it all is pretty bad even by 1990 standards.